+++ NEWS.md
@@ -1,3 +1,9 @@
+### cholera 0.8.0.9395
+
+- add/set euclideanPath(square.intersections = FALSE); amend location argument.
+- amend casePumpEucl().
+
+
 ### cholera 0.8.0.9394
 
 - fix typo in validateCase().